requirements.txt
adium-theme-ubuntu==0.3.4
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.5.1
bleach==1.5.0
boto==2.37.0
certifi==2017.7.27.1
configparser==3.5.0
dargparse==0.2.5
decorator==4.1.2
Django==1.11.5
django-mongodb-engine==0.6.0
djangorestframework==3.6.3
djangotoolbox==1.8.0
entrypoints==0.2.2
enum34==1.1.6
et-xmlfile==1.0.1
eyeD3==0.7.10
funcsigs==1.0.2
functools32==3.2.3.post2
gunicorn==19.6.0
gyp==0.1
h5py==2.6.0
hmmlearn==0.2.0
html5lib==0.9999999
importlib==1.0.4
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
jdcal==1.3
Jinja2==2.9.6
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.1
Keras==1.2.1
lxml==3.7.1
MarkupSafe==1.0
mistune==0.7.3
mock==2.0.0
nbconvert==5.1.1
nbformat==4.2.0
notebook==4.3.1
numexpr==2.6.1
numpy==1.13.1
openpyxl==2.4.1
pandas==0.19.1
pandocfilters==1.4.1
pathlib2==2.2.1
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
progressbar==2.3
prompt-toolkit==1.0.9
protobuf==3.1.0.post1
psutil==1.2.1
ptyprocess==0.5.2
pydub==0.16.7
Pygments==2.2.0
pymongo==2.9.5
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.2.1
requests==2.12.4
scandir==1.4
schedule==0.4.3
scikit-learn==0.18.1
scipy==0.19.1
selenium==3.0.2
simplegeneric==0.8.1
simplejson==3.10.0
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
sox==1.2.5
tables==3.3.0
tensorflow==0.12.1
terminado==0.6
testpath==0.3
Theano==0.8.2
tornado==4.5.2
traitlets==4.3.1
ubuntuone-client-data==14.4
unity-lens-photos==1.0
verlib==0.1
virtualenv==15.1.0
wcwidth==0.1.7
widgetsnbextension==1.2.6
yolog==0.3
youtube-dl==2017.1.25
